+++ src/Commands/Release/Wait.php
@@ -4,12 +4,12 @@ namespace SilverStripe\Cow\Commands\Release;
 
 use Exception;
 use SilverStripe\Cow\Model\Release\LibraryRelease;
-use SilverStripe\Cow\Steps\Release\Wait as WaitStep;
+use SilverStripe\Cow\Steps\Release\WaitStep;
 
 /**
  * Top level publish command
  */
-class Wait extends Release
+class Wait extends Publish
 {
     protected $name = 'release:wait';
 
@@ -24,18 +24,5 @@ class Wait extends Release
         // Once pushed, wait until installable
         $wait = new WaitStep($this, $project, $releasePlan);
         $wait->run($this->input, $this->output);
-    }
-
-    /**
-     * @return LibraryRelease
-     * @throws Exception
-     */
-    protected function getReleasePlan()
-    {
-        $plan = $this->getProject()->loadCachedPlan();
-        if (empty($plan)) {
-            throw new Exception("Please run 'cow release' before 'cow release:wait'");
-        }
-        return $plan;
     }
 }
